+++ src/pl/plajer/villagedefense3/utils/SetupInventory.java
@@ -1,6 +1,6 @@
 /*
  * Village Defense 3 - Protect villagers from hordes of zombies
- * Copyright (C) 2018  Plajer
+ * Copyright (C) 2018  Plajer's Lair - maintained by Plajer
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -134,6 +134,8 @@ public class SetupInventory {
 
     private String isOptionDoneList(String path) {
         if(ConfigurationManager.getConfig("arenas").isSet(path)) {
+            if(path.contains(".doors"))
+                return ChatColor.GOLD + "" + ChatColor.BOLD + "Done: " + ChatColor.GREEN + "Yes " + ChatColor.GRAY + "(value: " + ConfigurationManager.getConfig("arenas").getConfigurationSection(path).getKeys(false).size() / 2 + ")";
             return ChatColor.GOLD + "" + ChatColor.BOLD + "Done: " + ChatColor.GREEN + "Yes " + ChatColor.GRAY + "(value: " + ConfigurationManager.getConfig("arenas").getConfigurationSection(path).getKeys(false).size() + ")";
         }
         return ChatColor.GOLD + "" + ChatColor.BOLD + "Done: " + ChatColor.RED + "No";